+++ docs/versionhistory.rst
@@ -4,6 +4,9 @@ Version history
 This library adheres to
 `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
+**UNRELEASED**
+
+- Dropped Python 3.8 support
 - Changed the signature of ``typeguard_ignore()`` to be compatible with
   ``typing.no_type_check()`` (PR by @jolaf)